SeekBar输入输出改为float类型，可设置负数和步长

+++ dev-layout/src/main/java/com/github/eekidu/dev/devlayout/util/DevLayoutUtil.kt
@@ -72,7 +72,7 @@ fun DevLayout.addViewHeightAdjust(targetView: View) {
         targetView.layoutParams.height
     }
     this.addSeekBar("高度调节") { height ->
-        targetView.layoutParams.height = height
+        targetView.layoutParams.height = height.toInt()
         targetView.layoutParams = targetView.layoutParams
     }.setMax(2 * viewHeight).setProgress(viewHeight)
 }
